refactor(dsg): moving the input type enum to file
