+++ packages/amplication-data-service-generator/src/server/resource/dto/create-create-input.ts
@@ -2,7 +2,8 @@ import { builders, namedTypes } from "ast-types";
 import { Entity } from "../../../types";
 import { NamedClassDeclaration } from "../../../util/ast";
 import { isEditableField } from "../../../util/field";
-import { createInput, InputTypeEnum } from "./create-input";
+import { createInput } from "./create-input";
+import { InputTypeEnum } from "./input-type-enum";
 
 export function createCreateInput(entity: Entity): NamedClassDeclaration {
   const fields = entity.fields.filter(isEditableField);
