+++ packages/amplication-data-service-generator/src/server/resource/dto/input-type-enum.ts
@@ -1,0 +1,6 @@
+export enum InputTypeEnum {
+  Create = "create",
+  Update = "update",
+  Where = "where",
+  WhereUnique = "whereUnique",
+}
